pyproject.toml
[project]
name = "wisdomindata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=23.1.0",
    "asyncio>=3.4.3",
    "beautifulsoup4>=4.12.0",
    "bert-score>=0.3.13",
    "cryptography>=45.0.3",
    "datasets>=2.14.0",
    "evaluate>=0.4.4",
    "flask>=3.1.0",
    "fsspec>=2024.12.0",
    "gunicorn>=23.0.0",
    "httpx[http2]>=0.28.1",
    "huggingface-hub>=0.16.0",
    "ipython>=9.0.2",
    "kaleido==0.2.1",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.21",
    "langchain-neo4j>=0.4.0",
    "langchain-ollama>=0.3.1",
    "langchain-openai>=0.3.14",
    "langgraph>=0.3.29",
    "matplotlib>=3.7.0",
    "neo4j>=5.28.1",
    "networkx>=3.4.2",
    "nltk>=3.8.0",
    "numpy>=1.24.0",
    "ollama>=0.4.7",
    "openai>=1.70.0",
    "openpyxl>=3.1.5",
    "pandas>=1.5.0",
    "plotly>=5.15.0",
    "psycopg2-binary>=2.9.0",
    "pymysql>=1.1.0",
    "pytest>=8.3.5",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "rouge-score>=0.1.2",
    "sacrebleu>=2.3.0",
    "scikit-learn>=1.3.0",
    "seaborn>=0.12.0",
    "sqlalchemy>=2.0.0",
    "tabulate>=0.9.0",
    "torch>=2.0.0",
    "tqdm>=4.65.0",
    "transformers>=4.30.0",
    "vanna[chromadb]>=0.7.6",
    "xlrd>=2.0.1",
]
